Zoom::Tools: improve getScaledXFromValue() and getScaledYFromValue()

+++ Source/Zoom/AnlZoomTools.h
@@ -11,8 +11,8 @@ namespace Zoom
         Range getScaledVisibleRange(Accessor const& zoomAcsr, Range const& newGlobalRange);
         double getScaledValueFromWidth(Accessor const& zoomAcsr, juce::Component const& component, int x);
         double getScaledValueFromHeight(Accessor const& zoomAcsr, juce::Component const& component, int y);
-        int getScaledXFromValue(Accessor const& zoomAcsr, juce::Component const& component, double value);
-        int getScaledYFromValue(Accessor const& zoomAcsr, juce::Component const& component, double value);
+        double getScaledXFromValue(Accessor const& zoomAcsr, juce::Component const& component, double value);
+        double getScaledYFromValue(Accessor const& zoomAcsr, juce::Component const& component, double value);
     }
 }
